Refactor Main Page for easier changes
